+++ spec/polyfill/v2_4/integer/instance/ceil_spec.rb
@@ -29,5 +29,9 @@ RSpec.describe 'Integer#ceil' do
       expect(value).to receive(:to_int).and_return(1)
       1.ceil(value)
     end
+
+    it 'raises a TypeError when given a non-Integer' do
+      expect { 1.ceil('a') }.to raise_error TypeError
+    end
   end
 end