+++ spec/polyfill/v2_4/integer/instance/truncate_spec.rb
@@ -30,5 +30,9 @@ RSpec.describe 'Integer#truncate' do
       expect(value).to receive(:to_int).and_return(1)
       1.truncate(value)
     end
+
+    it 'raises a TypeError when given a non-Integer' do
+      expect { 1.truncate('a') }.to raise_error TypeError
+    end
   end
 end